requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1